Crate rusoto_comprehendmedical

Comprehend Medical extracts structured information from unstructured clinical text. Use these actions to gain insight in your documents.

If you're using the service, you're probably looking for ComprehendMedicalClient and ComprehendMedical.

Structs

Attribute
An extracted segment of the text that is an attribute of an entity, or otherwise related to an entity, such as the dosage of a medication taken. It contains information about the attribute such as id, begin and end offset within the input text, and the segment of the input text.

ComprehendMedicalClient
A client for the ComprehendMedical API.

DetectEntitiesRequest

DetectEntitiesResponse

DetectPHIRequest

DetectPHIResponse

Entity
Provides information about an extracted medical entity.

Trait
Provides contextual information about the extracted entity.

UnmappedAttribute
An attribute that we extracted, but were unable to relate to an entity.

Enums

DetectEntitiesError
Errors returned by DetectEntities

DetectPHIError
Errors returned by DetectPHI

Traits

ComprehendMedical
Trait representing the capabilities of the ComprehendMedical API. ComprehendMedical clients implement this trait.